api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
supabase==1.0.4
pytest==7.4.2
httpx==0.24.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.0.1
pytesseract==0.3.10
google-cloud-vision==3.4.3
openai==0.28.0
numpy==1.26.0
scikit-learn==1.3.0 